Prepare for the 2.10 Dart SDK (#29)
Set the min SDK to 2.10.0-0 to get a language version of `2.10`.

Move dependency overrides to the master branches for the packages which
support the 2.10 SDK on the master branch.

Update travis config to run with `--enable-experiment=non-nullable` and
test on the `null_safety` branch. Use the pattern from `package:collection`.

Delete some migration comments that are left over.